Switch to gulp from the Makefile, move static=>public
This makes the dev environment consistent with deploy, which
should help avoid issues with new files being needing to be
added to the deploy script in the future.

gulp also provides a simpler, more consistent way to describe
transformation steps. This will make it easier in the future
to add more steps (e.g. compile/concat/minify CSS).

Also renamed static to public, which is a bit clearer.